requirements.txt
# SocialProphet Dependencies
# Install with: pip install -r requirements.txt

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Time-Series Forecasting
prophet>=1.1.4
statsmodels>=0.14.0
pmdarima>=2.0.3  # auto_arima

# Machine Learning
scikit-learn>=1.3.0

# Deep Learning (Optional - for LSTM)
# tensorflow>=2.13.0
# torch>=2.0.0

# LLM & NLP
transformers>=4.35.0
requests>=2.31.0
huggingface-hub>=0.19.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Dashboard
streamlit>=1.28.0

# Data Collection
tweepy>=4.14.0  # Twitter API

# Text Processing
nltk>=3.8.0
textblob>=0.17.1

# Environment & Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
